Added util and validated user using zod

+++ Practice_1/MongoDB_Basics/Controller/MongoDB_Controller/UserActivityController.js
@@ -1,7 +1,18 @@
 const UserService = require("../../Services/UserService");
+const { validateUserInput, validateUserUpdate } = require("../../Validator/UserInputValidation");
+const isValidObjectId = require("../../utils/validateObjectId");
 
 // CREATE
 const createUser = async (req, res) => {
+    // Validating User sent data
+    const result = validateUserInput(req.body);
+    if (!result.success) {
+        return res.status(400).json({
+            error: result.error.issues.map(issue => issue.message)
+        });
+    }
+
+    // If Input good then create user in DB
     try {
         const user = await UserService.createUser(req.body);
         res.status(201).json(user);
@@ -22,6 +33,12 @@ const getAllUsers = async (req, res) => {
 
 // READ one
 const getUserById = async (req, res) => {
+    // checks the data before returning response
+    if (!isValidObjectId(req.params.id)) {
+        return res.status(400).json({ error: "Invalid user ID format." });
+    }
+
+    // If Input good then read user from DB
     try {
         const user = await UserService.getUserById(req.params.id);
         if (!user) return res.status(404).json({ error: "User not found" });
@@ -33,6 +50,22 @@ const getUserById = async (req, res) => {
 
 // UPDATE
 const updateUser = async (req, res) => {
+    if (!isValidObjectId(req.params.id)) {
+        return res.status(400).json({ error: "Invalid user ID format." });
+    }
+
+    const result = validateUserUpdate(req.body);
+    if (!result.success) {
+        return res.status(400).json({
+            error: result.error.issues.map(issue => issue.message)
+        });
+    }
+
+    if (Object.keys(result.data).length === 0) {
+        return res.status(400).json({ error: "At least one field must be provided to update." });
+    }
+
+    // If input is good then update the user
     try {
         const user = await UserService.updateUser(req.params.id, req.body);
         if (!user) return res.status(404).json({ error: "User not found" });
@@ -44,6 +77,11 @@ const updateUser = async (req, res) => {
 
 // DELETE
 const deleteUser = async (req, res) => {
+    if (!isValidObjectId(req.params.id)) {
+        return res.status(400).json({ error: "Invalid user ID format." });
+    }
+
+    // If Input is good then connect to DB to delete the user
     try {
         const user = await UserService.deleteUser(req.params.id);
         if (!user) return res.status(404).json({ error: "User not found" });
